+++ forui/lib/src/widgets/calendar/month/month_picker.dart
@@ -69,6 +69,7 @@ class _MonthPickerState extends State<MonthPicker> {
   Widget build(BuildContext context) => Padding(
         padding: const EdgeInsets.only(top: 5.0),
         child: GridView(
+          padding: EdgeInsets.zero,
           gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
             crossAxisCount: YearPicker.columns,
             mainAxisExtent: ((DayPicker.tileDimension - 5.0) * DayPicker.maxRows) / YearPicker.rows,