+++ forui/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 * Add `FButtonCustomStyle.enabledHoverBoxDecoration`.
 
+### Changes
+
+* Change `FCalendar` year  picker to update the header whenever a year is selected.
+* Increase `FCalendar`'s default text size from `FTypography.sm` to `FTypography.base`.
+* **Breaking** Split `FCalendarHeaderStyle.iconColor` into `FCalendarHeaderStyle.enabledIconColor` and
+  `FCalendarHeaderStyle.disabledIconColor`.
+
+### Fixes
+
+* Fix `FBottomNavigationBar` items hit region being smaller than intended.
+* Fix `FCalendar` showing focused outline when pressing and long pressing a date.
+* Fix `FCalendar` year and month picker applying incorrect initial top padding.
+* Fix `FCalendar` year and month picker incorrectly calculating start and end dates.
+
 
 ### Changes